Adds two additional classes and changes the logic
