+++ permissions.rb
@@ -1,0 +1,16 @@
+class Permissions
+  def permission?
+    print 'Has parent permission? [Y/N]:'
+    permission = gets.chomp
+
+    case permission
+    when 'n', 'N'
+      false
+    when 'y', 'Y'
+      true
+    else
+      puts 'Incorrect choice, kindly enter \'y\', \'Y\' or \'n\', \'N\' '
+      permission?
+    end
+  end
+end
